+++ CHANGELOG.md
@@ -13,15 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a new NLP metric `InfoLM` ([#915](https://github.com/PyTorchLightning/metrics/pull/915))
 
-
 - Added `Perplexity` metric ([#922](https://github.com/PyTorchLightning/metrics/pull/922))
-
 
 - Added argument `normalize` to `LPIPS` metric ([#1216](https://github.com/Lightning-AI/metrics/pull/1216))
 
 - Added support for multiprocessing of batches in `PESQ` metric ([#1227](https://github.com/Lightning-AI/metrics/pull/1227))
 
 - Added support for multioutput in `PearsonCorrCoef` and `SpearmanCorrCoef` ([#1200](https://github.com/Lightning-AI/metrics/pull/1200))
+
+- Added a new segmentation metric `mean IoU` ([#915](https://github.com/PyTorchLightning/metrics/pull/915))
 
 ### Changed